updated shopping cart functionality: improved cart item quantity updates, added validation, and adjusted transactional annotations

+++ src/main/java/mate/academy/service/ShoppingCartService.java
@@ -12,7 +12,7 @@ public interface ShoppingCartService {
 
     ShoppingCartDto addCartItemToCart(Long userId, CreateCartItemRequestDto requestDto);
 
-    ShoppingCartDto updateCartItemQuantity(Long userId, Long cartItemId, Long quantity);
+    ShoppingCartDto updateCartItemQuantity(Long userId, Long cartItemId, Integer quantity);
 
     void removeCartItemFromCart(Long userId, Long cartItemId);
 }
